refactor: PostsManager 관련 컴포넌트 분리 및 feature 디렉토리 index 정리
