+++ src/feature/comment/hooks/useCommentDialogState.ts
@@ -1,0 +1,11 @@
+import { useToggle } from "../../../shared/hooks/useToggle"
+
+export default function useCommentDialogState() {
+  const showAddCommentDialog = useToggle(false)
+  const showEditCommentDialog = useToggle(false)
+
+  return {
+    showAddCommentDialog,
+    showEditCommentDialog,
+  }
+}
