+++ src/feature/comment/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as CommentAddDialog } from "./CommentAddDialog"
+export { default as CommentEditDialog } from "./CommentEditDialog"
