+++ src/feature/post/hooks/usePostDialogState.ts
@@ -1,0 +1,13 @@
+import { useToggle } from "../../../shared/hooks/useToggle"
+
+export const usePostDialogState = () => {
+  const showAddDialog = useToggle(false)
+  const showEditDialog = useToggle(false)
+  const showPostDetailDialog = useToggle(false)
+
+  return {
+    showAddDialog,
+    showEditDialog,
+    showPostDetailDialog,
+  }
+}
