+++ src/feature/post/ui/index.ts
@@ -1,0 +1,6 @@
+export { default as PostTable } from "./PostTable"
+export { default as PostSearchFilter } from "./PostSearchFilter"
+export { default as PostPageNation } from "./PostPageNation"
+export { default as PostAddDialog } from "./PostAddDialog"
+export { default as PostEditDialog } from "./PostEditDialog"
+export { default as PostDetailDialog } from "./PostDetailDialog"
